pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "trip-planner"
version = "0.1.0"
description = "Luxury Europe Trip Planner — Telegram Bot Backend"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    # Pydantic v2
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Queue / Cache
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    # Telegram bot
    "python-telegram-bot>=20.7",
    # Payment
    "stripe>=9.0.0",
    # Storage
    "boto3>=1.34.0",
    # Document generation
    "reportlab>=4.1.0",
    "python-docx>=1.1.0",
    # Observability
    "structlog>=24.1.0",
    "sentry-sdk[fastapi]>=2.5.0",
    # HTTP client (bot → API calls)
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
